evals: share the observe selector-identity check
Six ported observe tasks each carried the same in-page resolver for comparing
an observed selector against expected ones. Selector strings can't be compared
textually — xpath=//button[@id="go"], #go and /html/body/div/button may all
address one element — so each task resolved both sides inside the page and
compared element references.

The resolver could not simply be imported: page.evaluate serializes its
callback and runs it in the browser, where module imports do not exist. So the
whole evaluate call moves into framework/observeSelectors.ts, which inlines the
resolver once and returns the matching selector (null if none match). Four
tasks use that selector; the two that only need a yes/no compare against null.

The extracted resolver is byte-identical to the six copies it replaces, and
deliberately omits the try/catch a variant on the harness branch grew: that
would turn a malformed-selector throw into a silent non-match, converting an
error row into a failed one.
