+++ packages/evals/tasks/bench/observe/observe_vantechjournal.ts
@@ -1,4 +1,5 @@
 import { defineBenchV4Task } from "../../../framework/defineTask.js";
+import { matchingSelector } from "../../../framework/observeSelectors.js";
 
 export default defineBenchV4Task(
   { name: "observe_vantechjournal" },
@@ -29,41 +30,8 @@ export default defineBenchV4Task(
       // re-expressed in-page: resolve the observed selector and each
       // expected selector and compare element references. Expected locators
       // that fail to resolve are skipped, as in v3.
-      const foundMatch = await page.evaluate(
-        ({
-          observedSelector,
-          expectedSelectors,
-        }: {
-          observedSelector: string;
-          expectedSelectors: string[];
-        }) => {
-          const resolve = (selector: string): Element | null => {
-            const raw = selector.startsWith("xpath=") ? selector.slice("xpath=".length) : selector;
-            if (raw.startsWith("/") || raw.startsWith("(")) {
-              const result = document.evaluate(
-                raw,
-                document,
-                null,
-                XPathResult.FIRST_ORDERED_NODE_TYPE,
-                null,
-              );
-              return result.singleNodeValue as Element | null;
-            }
-            return document.querySelector(raw);
-          };
-
-          const observed = resolve(observedSelector);
-          if (!observed) return false;
-          for (const expected of expectedSelectors) {
-            if (resolve(expected) === observed) return true;
-          }
-          return false;
-        },
-        {
-          observedSelector: observations[0].selector,
-          expectedSelectors: expectedLocators,
-        },
-      );
+      const foundMatch =
+        (await matchingSelector(page, observations[0].selector, expectedLocators)) !== null;
 
       return {
         _success: foundMatch,
